Reworked tests structure
